Factor addPopup
